SoilMoisture

Documentation for SoilMoisture.

SoilMoisture.dt2daily

SoilMoisture.evapotranspiration

SoilMoisture.leakage

SoilMoisture.rainfall_poisson

SoilMoisture.soil_moisture_pdf

SoilMoisture.soil_rp

SoilMoisture.soil_water_balance

SoilMoisture.solve_swb

SoilMoisture.water_loss

SoilMoisture.dt2daily — Method

Convert soil water balance results to daily time-scale

SoilMoisture.evapotranspiration — Method

evapotranspirationi(s, sh, sw, sstar, emax, ew)

Computes evapotranspiration losses

Arguments

s: soil moisture

sh: soil moisture at hygroscopic point.

sw: soil moisture at wilting point.

sstar: soil moisture below field capacity.

emax: maximum evapotranspiration rate.

ew: soil evaporation rate

Example

s = 0.45
sh = 0.12
sw = 0.20
sstar = 0.55
emax = 0.50
ew = 0.05

et = evapotranspiration(s, sh, sw, sstar, emax, ew)

SoilMoisture.leakage — Method

leakage(s, sfc, β, ks)

Computes the soil leakage

Arguments

s: soil moisture.

sfc: soil moisture at field capacity.

b: leakage curve exponent.

ks: hydralic conductivity.

Example

s = 0.60
sfc = 0.55
β = 12.5
ks = 150

lk = leakage(s, sfc, β, ks)

SoilMoisture.rainfall_poisson — Method

rainfall_poisson(n, α, λ)

Generates rainfall using Poisson process.

Arguments

n::Integer: number of events.

α::Float64: mean rain depth.

λ::Float64: rain event interval (1/d).

If rainfall needs to be similated in timescale smaller than daily, multiply λ by dt.

Example

nevents = 30
α = 0.75
λ = 0.45
rain = rainfall_poisson(n, α, λ)

SoilMoisture.soil_moisture_pdf — Method

Soil moisture pdf

soil_moisture_pdf(s, sh, sw, sstar, sfc, β, ks, n, zr, ew, emax, α, λ)

Arguments

s: soil moisture.

sh: soil moisture at hygroscopic point.

sw: soil moisture at wilting point.

sstar: soil moisture below field capacity.

sfc: soil moisture at field capacity.

β: leakage curve exponent.

ks: hydralic conductivity.

n: porosity.

zr: root zone depth.

emax: maximum evapotranspiration rate.

ew: soil evaporation rate.

α: mean rainfall depth.

λ: mean rainfall frequency.

Example

s = [0.1:0.01:1;]
sh = 0.1
sw = 0.2
sstar = 0.45
sfc = 0.65
β = 12.5
ks = 120.0
n = 0.45
zr = 40.0
ew = 0.05
emax = 0.5
α = 0.95
λ = 0.5
p = soil_moisture_pdf(s, sh, sw, sstar, sfc, β, ks, n, zr, ew, emax, α, λ)

SoilMoisture.soil_rp — Method

Soil penetration resistance

soil_rp(s, bulk_density, a, b, c)

Arguments

s: soil moisture.

bulk_density: soil bulk density.

a, b, and c: soil penetration resistance parameters.

Example

s = 0.45
bd = 1.68
a = -5.75
b = 6.45
c = -15.30
rp = soil_rp(s, bulk_density, a, b, c)

SoilMoisture.soil_water_balance — Method

Computes the soil water balance

This function assumes that the canopy interception is negligible.

Arguments

rain: rainfall.

s: soil moisture.

sh: soil moisture at hygroscopic point.

sw: soil moisture at wilting point.

sstar: soil moisture below field capacity.

sfc: soil moisture at field capacity.

β: leakage curve exponent.

ks: hydralic conductivity.

n: porosity.

zr: root zone depth.

emax: maximum evapotranspiration rate.

ew: soil evaporation rate.

dt: time step.

SoilMoisture.solve_swb — Method

Solve soil water balance model

Arguments

rain: a vector of rainfall series at the same time step the model will be solved.

s: initial guess for the soil moisture value.

params: a dictionary with all parameters needed to for soil_water_balance() function.

SoilMoisture.water_loss — Method

water_loss(s, sh, sw, sstar, sfc, emax, ew, β, ks)

Water loss function

Combines evapotranspiration and leakage as a function of soil moisture.

Arguments

s: soil moisture.

sh: soil moisture at hygroscopic point.

sw: soil moisture at wilting point.

sstar: soil moisture below field capacity.

sfc: soil moisture at field capacity.

emax: maximum evapotranspiration rate.

ew: soil evaporation rate.

β: leakage curve exponent.

ks: hydralic conductivity.